Performance optimisations for renderer

+++ internal-renderers/src/d2-renderer/D2Renderer.ts
@@ -1,4 +1,4 @@
-import { map, throttle, times } from "lodash";
+import { debounce, find, isEqual, map, once, throttle, times } from "lodash";
 import { nanoid } from "nanoid";
 import { Viewport } from "pixi-viewport";
 import * as PIXI from "pixi.js";
@@ -11,10 +11,12 @@ import {
   D2RendererOptions,
   defaultD2RendererOptions,
 } from "./D2RendererOptions";
-import { D2WorkerEvent } from "./D2RendererWorker";
+import { D2WorkerEvent, getTiles } from "./D2RendererWorker";
 import { D2RendererWorkerAdapter } from "./D2RendererWorkerAdapter";
 import { EventEmitter } from "./EventEmitter";
 import { intersect } from "./intersect";
+
+const { max } = Math;
 
 class Tile extends PIXI.Sprite {
   constructor(texture?: PIXI.Texture, public bounds?: Bounds) {
@@ -29,6 +31,7 @@ class D2Renderer
   #app?: PIXI.Application<HTMLCanvasElement>;
   #viewport?: Viewport;
   #world?: PIXI.Container<Tile>;
+  #grid?: PIXI.Graphics;
   #options: D2RendererOptions = defaultD2RendererOptions;
   #workers: D2RendererWorkerAdapter[] = [];
 
@@ -45,6 +48,7 @@ class D2Renderer
 
   destroy(): void {
     map(this.#workers, (w) => w.terminate());
+    this.#app!.destroy();
   }
 
   add(components: CompiledD2IntrinsicComponent[]) {
@@ -56,13 +60,14 @@ class D2Renderer
   setOptions(options: D2RendererOptions) {
     this.#handleWorkerChange(options);
     this.#handleWindowSizeChange(options);
-    this.#handleFrustumChange();
     this.#options = options;
+    this.#handleFrustumChange();
+    this.#updateGrid();
   }
 
   #setupPixi(options: D2RendererOptions) {
     this.#app = new PIXI.Application({
-      background: "#ffffff",
+      backgroundAlpha: 0,
       width: options.screenSize.width,
       height: options.screenSize.height,
     });
@@ -84,17 +89,25 @@ class D2Renderer
       .decelerate({ friction: 0.98 })
       .clampZoom({ maxScale: 300, minScale: 0.01 });
 
-    this.#viewport.on(
-      "moved",
-      throttle(() => this.#handleFrustumChange(), options.refreshInterval, {
-        leading: false,
-        trailing: true,
-      })
-    );
+    this.#viewport.on("moved", () => {
+      this.#getFrustumChangeQueue()();
+      this.#getUpdateGridQueue()();
+    });
 
     this.#world = new PIXI.Container();
     this.#viewport.addChild(this.#world);
+
+    this.#grid = new PIXI.Graphics();
+    this.#viewport.addChild(this.#grid);
   }
+
+  #getFrustumChangeQueue = once(() =>
+    debounce(() => this.#handleFrustumChange(), this.#options.debounceInterval)
+  );
+
+  #getUpdateGridQueue = once(() =>
+    debounce(() => this.#updateGrid(), this.#options.refreshInterval)
+  );
 
   #handleWorkerChange(options: D2RendererOptions) {
     map(this.#workers, (w) => w.terminate());
@@ -129,24 +142,72 @@ class D2Renderer
     );
   }
 
-  #addToWorld(texture: PIXI.Texture, bounds: Bounds) {
-    const { tileResolution: resolution } = this.#options;
-    const scale = {
-      x: (bounds.right - bounds.left) / resolution.width,
-      y: (bounds.bottom - bounds.top) / resolution.height,
-    };
-    this.#cull(bounds);
-    this.#world
-      ?.addChild(new Tile(texture, bounds))
-      .setTransform(bounds.left, bounds.top, scale.x, scale.y);
-  }
-
-  #cull(bounds: Bounds) {
-    for (const c of this.#world!.children) {
-      if (intersect(c.bounds!, bounds)) {
-        c.destroy({ texture: true, baseTexture: true });
+  #updateGrid() {
+    const { tileSubdivision, accentColor } = this.#options;
+    const { zoom, tiles } = getTiles(this.#viewport!, tileSubdivision);
+    const px = this.#getPx();
+    this.#grid?.clear();
+    this.#grid?.removeChildren();
+    this.#grid?.lineStyle(1 * px, accentColor, 0.5);
+    this.#grid?.beginFill(accentColor, 0.05);
+    for (const { bounds: b, tile: t } of tiles) {
+      if (!find(this.#world?.children, (c) => isEqual(c.bounds, b))) {
+        this.#grid?.drawRect(b.left, b.top, b.right - b.left, b.bottom - b.top);
+        const text = new PIXI.Text(
+          `LOD ${zoom} / X=${t.x} Y=${t.y} / L=${b.left} T=${b.top} R=${b.right} B=${b.bottom}`,
+          {
+            fontFamily: "Inter",
+            fontSize: 12 * px,
+            fill: accentColor,
+          }
+        );
+        text.resolution = devicePixelRatio / px;
+        this.#grid
+          ?.addChild(text)
+          .setTransform(b.left + 12 * px, b.top + 12 * px);
       }
     }
+  }
+
+  #getPx() {
+    const { right, left } = this.#viewport!;
+    const { width } = this.#options.screenSize;
+    return (right - left) / width;
+  }
+
+  async #addToWorld(texture: PIXI.Texture, bounds: Bounds) {
+    const { tileResolution: resolution, tileSubdivision } = this.#options;
+    const { tiles } = getTiles(this.#viewport!, tileSubdivision);
+    if (find(tiles, (t) => isEqual(t.bounds, bounds))) {
+      const scale = {
+        x: (bounds.right - bounds.left) / resolution.width,
+        y: (bounds.bottom - bounds.top) / resolution.height,
+      };
+      const tile = new Tile(texture, bounds);
+      this.#world
+        ?.addChild(tile)
+        .setTransform(bounds.left, bounds.top, scale.x, scale.y);
+      this.#getUpdateGridQueue()();
+      await this.#animate(tile);
+      for (const c of this.#world!.children) {
+        if (intersect(c.bounds!, bounds) && c !== tile) {
+          c.destroy({ texture: true, baseTexture: true });
+        }
+      }
+    }
+  }
+
+  #animate(tile: Tile) {
+    return new Promise<void>((res) => {
+      const f = () => {
+        if (tile.alpha < 1) {
+          tile.alpha += 1000 / 60 / this.#options.animationDuration;
+          requestAnimationFrame(f);
+        } else res();
+      };
+      tile.alpha = 0;
+      requestAnimationFrame(f);
+    });
   }
 }
